Simplifying View classes
